Move getLanguageSelfName to tools.h
This is a general utility which other ports can get use of.
Added tests
